+++ src/components/PersonalFormFields/index.tsx
@@ -16,10 +16,7 @@ interface PersonalFormFieldsProps {
   agencyType?: string;
   id?: string;
 }
-function PersonalFormFields({
-  agencyType = "",
-  id = "",
-}: PersonalFormFieldsProps) {
+function PersonalFormFields({ id = "" }: PersonalFormFieldsProps) {
   const { t } = useTranslation("common");
   const {
     register,
@@ -61,7 +58,7 @@ function PersonalFormFields({
       </FormRow>
       <FormRow id={`${id}-personalForm-2`}>
         <DSFormField>
-          <AgeFormFields policyType={agencyType || formValues.policyType} />
+          <AgeFormFields />
         </DSFormField>
       </FormRow>
       <FormRow id={`${id}-personalForm-3`}>